+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [2.19.0] 2017-09-06
+
+### Added
+- 'ScrollingMixin' mixin to scroll elements into the visible viewport of scrollable areas
+- provide `<BreadcrumbList/>` and `<BreadcrumbItem/>` to create breadcrumb navigations
+
 ## [2.18.2] 2017-08-31
 ### Added
 - `<ProgressButton />` event `progressTopic` can now additionally contain param `lastUpdate`.